+++ bin/run_pytests
@@ -11,4 +11,7 @@ mysql -uroot -proot -e "DROP DATABASE IF EXISTS ggrcdevtest; CREATE DATABASE ggr
 export GGRC_SETTINGS_MODULE="testing ggrc_basic_permissions.settings.development ggrc_gdrive_integration.settings.development ggrc_risk_assessments.settings.development ggrc_risks.settings.development ggrc_workflows.settings.development"
 db_migrate
 
+echo -e "\nRunning unit tests"
+nosetests unit --logging-clear-handlers ${@:1}
+echo -e "\nRunning integration tests"
 nosetests integration --logging-clear-handlers ${@:1}
